+++ niryo_robot_system_api_client/src/niryo_robot_system_api_client/SystemApiEndpoint.py
@@ -16,3 +16,4 @@ class SystemApiEndpoint(Enum):
     ETHERNET_PROFILE = '/ethernetProfile'
     SET_SETTING = '/setSetting'
     GET_SETTING = '/getSetting'
+    SYSTEM_VERSION_CURRENT = '/system/version/current'